FMI export utilities work with Visual Studio (incl. tests), also some cosmetic changes to silence MSVC warnings

+++ test/zigzag_fmu/zigzag.c
@@ -75,7 +75,7 @@ DllExport fmiStatus fmiSetTime( fmiComponent c, fmiReal time )
 DllExport fmiStatus fmiSetContinuousStates( fmiComponent c, const fmiReal x[], size_t nx )
 {
 	fmustruct* fmu = (fmustruct*) c;
-	int i;
+	size_t i;
 	for ( i = 0; i < nx; i++ )
 		fmu->rvar[i] = x[i];
 
@@ -105,7 +105,7 @@ DllExport fmiStatus fmiCompletedIntegratorStep( fmiComponent c, fmiBoolean* call
 DllExport fmiStatus fmiSetReal( fmiComponent c, const fmiValueReference vr[], size_t nvr, const fmiReal value[] )
 {
 	fmustruct* fmu = (fmustruct*) c;
-	int i;
+	size_t i;
 	for ( i = 0; i < nvr; i++ )
 		fmu->rvar[vr[i]] = value[i];
 
@@ -182,7 +182,7 @@ DllExport fmiStatus fmiGetEventIndicators( fmiComponent c, fmiReal eventIndicato
 DllExport fmiStatus fmiGetReal( fmiComponent c, const fmiValueReference vr[], size_t nvr, fmiReal value[] )
 {
 	fmustruct* fmu = (fmustruct*) c;
-	int i;
+	size_t i;
 	for ( i = 0; i < nvr; i++ )
 		value[i] = fmu->rvar[vr[i]];
 
@@ -225,7 +225,7 @@ DllExport fmiStatus fmiEventUpdate( fmiComponent c, fmiBoolean intermediateResul
 DllExport fmiStatus fmiGetContinuousStates( fmiComponent c, fmiReal states[], size_t nx )
 {
 	fmustruct* fmu = (fmustruct*) c;
-	int i;
+	size_t i;
 	for( i = 0; i < nx; i++ )
 		states[i] = fmu->rvar[i];
 
@@ -242,7 +242,7 @@ DllExport fmiStatus fmiGetNominalContinuousStates( fmiComponent c, fmiReal x_nom
 DllExport fmiStatus fmiGetStateValueReferences( fmiComponent c, fmiValueReference vrx[], size_t nx )
 {
 	fmustruct* fmu = (fmustruct*) c;
-	int i;
+	size_t i;
 	for( i = 0; i < nx; i++ )
 		vrx[i] = fmiUndefinedValueReference;
